pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snowflake-cli-devops"
version = "0.1.0"
description = "A Snowflake CLI plugin for DevOps"
requires-python = ">=3.10"
dependencies = [
    "snowflake-cli>=3.0.0"
]

[project.entry-points."snowflake.cli.plugin.command"]
devops = "snowflake_cli_devops.plugin_spec"

[tool.setuptools.packages.find]
where = ["src"]
